+++ changelog.md
@@ -1,4 +1,9 @@
 
+New Feature:
+
++ FS::unmount()/FS::remount()
+
+
 # v1.0.6 2026/2/3
 
 New Feature: